Move signer helpers into src/nostr/event.js (#341 round-3)
Two Copilot points addressed:

- Repo-root scripts (clock-updater.mjs, test-nostr-auth.js,
  test-git-nostr-auth.js) previously imported from
  test/helpers/nostr-event.js, which was documented as test-only.
  That coupled non-test code to test-only code. Moved the signer
  surface (generateSecretKey, getPublicKey, finalizeEvent,
  nip98Token) into src/nostr/event.js alongside the verifier surface
  it depends on. test/helpers/ is now empty and removed.
- generateSecretKey: dropped the `crypto.randomBytes(32)` fallback.
  A bare 32-byte random can produce 0 or a value >= curve order — a
  vanishingly small but nonzero probability of an invalid secp256k1
  scalar. Now we throw if `secp256k1.utils.randomPrivateKey` is
  unavailable rather than returning a possibly-invalid key.

All importers updated:
- test/nostr-event.test.js + test/did-nostr.test.js: import from
  ../src/nostr/event.js
- clock-updater.mjs, test-nostr-auth.js, test-git-nostr-auth.js:
  import from ./src/nostr/event.js

Full suite: 523/523.
